requirements.txt

pandas
scikit-learn	
mlflow	
fastapi	
uvicorn	
dvc
evidently	
pytest	
pyyaml	
flake8	